Header, cpp, driver
I have done the declaration of the classes in the header file and the implementation of the classes OrdersList, Order and Deploy, except for the stream insertion operator. These are tested in the driver file.
TODO: Implementation subclasses of Order: Airlift, Blockade, Advance, Bomb, Negotiate. Declaration and implementation of stream insertion operator for all the classes.
Note: The declaration and implementation of classes Territory and Player are for testing purposes. These are dummy classes.